requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
groq==0.9.0
networkx==3.3
supabase==2.4.2
python-dotenv==1.0.1
pandas==2.2.2